requirements.txt
streamlit
openai-whisper
torch
sentence-transformers
chromadb
groq
pydub
tqdm